[OPP-1504] innstillinger-modal-form (#1970)
